requirements.txt
# Dependencies for Bili-Summarizer
# Last reviewed: 2025-12-25
google-generativeai
python-dotenv
yt-dlp
fastapi
uvicorn[standard]
python-multipart
jinja2
gunicorn
httpx
python-pptx
supabase
reportlab
python-alipay-sdk
cryptography
qrcode
pillow
psycopg2-binary
edge-tts
apscheduler
pywebpush
